+++ joulescope_ui/urls.py
@@ -1,0 +1,18 @@
+# Copyright 2023 Jetperch LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+UI_USERS_GUIDE = 'https://download.joulescope.com/products/joulescope_ui/users_guide/index.html'
+JS110_USERS_GUIDE = 'https://download.joulescope.com/docs/JoulescopeUsersGuide/index.html'
+JS220_USERS_GUIDE = 'https://download.joulescope.com/products/JS220/JS220-K000/users_guide/index.html'